Liberalize what can be used after ..
And fix the descent to be more robust, and avoid various crashes we
might get while trying to match too.

Also, it turns out that `..['foo']bar` can be accepted, without the
requirement of a `.` before the `bar`, in some implementations of
JSONPath, so allow that in ours too. It simplifies the parsing a
bit.